Inventory System / Loot System implemented

+++ Val'Hein/Assets/Scripts/Monobehaviours/General/IDamageable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-	void TakeDamage(float ammount);
+	void TakeDamage(int ammount);
 }
